ml-esports-platform/backend/requirements.txt
django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3.1
PyMySQL>=1.1.0
Pillow>=10.3
django-cors-headers>=4.3
gunicorn>=21.2
whitenoise>=6.6
python-dotenv>=1.0
django-cloudinary-storage>=0.3.0
cloudinary>=1.36.0
firebase-admin>=6.5.0
channels>=4.0.0
daphne>=4.0.0

